Fixing a module of unit tests so that they finish faster.
